Fixes and clean up following suggested changes
